% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_dl_data.R
\docType{data}
\name{cran_dl_data}
\alias{cran_dl_data}
\title{CRAN downloads for 20 tidyverse packages}
\format{
A data frame with 77940 rows and 3 variables:
\describe{
  \item{date}{The date}
  \item{count}{Number of downlaods in the date}
  \item{package}{Name of package}
}
}
\usage{
cran_dl_data
}
\description{
A dataset containing the daily number of downloads for the supplied packages
}
\keyword{datasets}
