% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_cran_data}
\alias{get_cran_data}
\title{Title}
\usage{
get_cran_data(package_names, from_date, to_date = Sys.Date())
}
\arguments{
\item{package_names}{A character vector, names of packages to download data for}

\item{from_date}{A string}

\item{to_date}{A string, default is set to today's date}
}
\value{
A \code{data.frame}
}
\description{
Title
}
\examples{
get_cran_data("rsconnect", from_date = (Sys.Date() - 10))

}
