% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cran_name}
\alias{get_cran_name}
\title{Get only package name from repo name}
\usage{
get_cran_name(repo)
}
\arguments{
\item{repo}{\code{string} The repo name, e.g., \bold{"rstudio/rsconnect"}}
}
\value{
\code{string} The package name, e.g., \bold{"rsconnect"}
}
\description{
Get only package name from repo name
}
\examples{
get_cran_name("rsconnect/rstudio")
}
