% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_get_data.R
\name{get_gh_stars}
\alias{get_gh_stars}
\title{Get the github stars for the supplied repo on a daily basis}
\usage{
get_gh_stars(repo)
}
\arguments{
\item{repo}{A \code{character} vector containing the repo name, eg. "rstudio/rsconnect"}
}
\value{
A \code{data.frame}
}
\description{
Get the github stars for the supplied repo on a daily basis
}
\examples{
get_gh_stars("rstudio/rsconnect")

}
