% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_plots.R
\name{gh_plot}
\alias{gh_plot}
\title{Title}
\usage{
gh_plot(r)
}
\arguments{
\item{r}{\code{reactiveValues} object with 4 slots, repo, date, cran_dl & gh_stars}
}
\value{
\code{ggplot}
}
\description{
Title
}
