% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_stars_data.R
\docType{data}
\name{gh_stars_data}
\alias{gh_stars_data}
\title{Github Stars for 20 tidyverse packages}
\format{
A data frame with 12748 rows and 3 variables:
\describe{
  \item{date}{The date}
  \item{n_stars}{Number of stars in the date}
  \item{package}{Repo of package}
}
}
\usage{
gh_stars_data
}
\description{
A dataset containing the daily number of stars the packages got
}
\keyword{datasets}
