% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\docType{data}
\name{repos}
\alias{repos}
\title{20 tidyverse packages}
\format{
A data frame with 20 rows and 1 variable:
\describe{
  \item{tidypacks}{The name of the github repo}
}
}
\usage{
repos
}
\description{
A dataset containing 20 tidyverse packages in the form of github repos.
}
\keyword{datasets}
