% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{theme_calendar}
\alias{theme_calendar}
\title{Theme for the github calendar plot}
\usage{
theme_calendar(base_size = 15, base_family = "sans")
}
\arguments{
\item{base_size}{\code{numeric} Base font size}

\item{base_family}{\code{string} Base font family}
}
\description{
Theme for the github calendar plot
}
