% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get-data.R
\name{get_cran_name}
\alias{get_cran_name}
\title{Get only package name from repo name}
\usage{
get_cran_name(repo)
}
\arguments{
\item{repo}{The repo name, e.g \code{"rstudio/rsconnect"}}
}
\value{
A string, the package name, e.g \code{"rsconnect"}
}
\description{
Get only package name from repo name
}
\examples{
get_cran_name("rsconnect/rstudio")
}
